Compiler warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/trunk@1092 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b

+++ vnc_unixsrc/Xvnc/programs/Xserver/hw/vnc/sprite.c
@@ -195,7 +195,7 @@ static GCOps rfbSpriteGCOps = {
     (rfbScreen.cursorIsDrawn &&					    \
         (pScreenPriv->pCacheWin == pWin ?			    \
 	    pScreenPriv->isInCacheWin : (			    \
-	    ((int) (pScreenPriv->pCacheWin = (pWin))) ,		    \
+	    ((long) (pScreenPriv->pCacheWin = (pWin))) ,		    \
 	    (pScreenPriv->isInCacheWin =			    \
 		(pWin)->drawable.x < pScreenPriv->saved.x2 &&	    \
 		pScreenPriv->saved.x1 < (pWin)->drawable.x +	    \
